I'm pushing up the database hierarchy diagram
